(PC-14598)[PRO] refactor: move type declaration into OfferEducational/core
