+++ pro/src/core/OfferEducational/types.ts
@@ -1,4 +1,9 @@
-import { OfferAddressType, OfferStatus } from 'api/v1/gen'
+import {
+  GetOfferVenueResponseModel,
+  OfferAddressType,
+  OfferStatus,
+  StudentLevels,
+} from 'api/v1/gen'
 
 export type IUserVenue = {
   id: string
@@ -159,4 +164,42 @@ export type CollectiveStockResponseModel = {
   isEducationalStockEditable?: boolean
   educationalPriceDetail?: string
   stockId?: string
+}
+
+export type CollectiveOfferResponseModel = {
+  bookingEmail?: string | null
+  dateCreated: Date
+  description?: string | null
+  durationMinutes?: number | null
+  students: StudentLevels[]
+  offerVenue: {
+    addressType: OfferAddressType
+    otherAddress: string
+    venueId: string
+  }
+  offerId?: string | null
+  contactEmail: string
+  contactPhone: string
+  hasBookingLimitDatetimesPassed: boolean
+  id: string
+  isActive: boolean
+  isBookable: boolean
+  audioDisabilityCompliant: boolean
+  mentalDisabilityCompliant: boolean
+  motorDisabilityCompliant: boolean
+  nonHumanizedId: number
+  visualDisabilityCompliant: boolean
+  name: string
+  collectiveStock: {
+    id: string
+    isBooked: boolean
+  }
+  subcategoryId: string
+  venue: GetOfferVenueResponseModel
+  venueId: string
+  status: OfferStatus
+}
+
+export type CollectiveOffer = CollectiveOfferResponseModel & {
+  isBooked: boolean
 }
