+++ pro/src/repository/pcapi/pcapi.ts
@@ -9,9 +9,9 @@ import {
 import {
   CollectiveStockResponseModel,
   CreateCollectiveStockPayload,
+  CollectiveOfferResponseModel,
 } from 'core/OfferEducational'
 import { ALL_OFFERERS, DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
-import { CollectiveOfferResponseModel } from 'core/Offers/types'
 import { client } from 'repository/pcapi/pcapiClient'
 import {
   FORMAT_ISO_DATE_ONLY,
